# Bekbook/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Mahalla Kutubxona API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reservations:
    post:
      operationId: createReservation
      tags: [reservations]
      summary: Reserve a book
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReservationBody"
      responses:
        "201":
          description: Reserved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "409":
          description: Already reserved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: getMyReservations
      tags: [reservations]
      summary: Get my reservations
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of reservations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationsListResponse"
...
  /reservations/{id}:
    delete:
      operationId: cancelReservation
      tags: [reservations]
      summary: Cancel a reservation
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Cancelled
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reservations/book/{bookId}:
    get:
      operationId: getBookReservationCount
      tags: [reservations]
      summary: Get reservation count for a book
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reservation count
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                required: [count]
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    Reservation:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        bookId:
          type: integer
          nullable: true
        storeBookId:
          type: integer
          nullable: true
        status:
          type: string
          enum: [waiting, ready, expired, cancelled]
        notifiedAt:
          type: string
          format: date-time
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        bookTitle:
          type: string
          nullable: true
      required: [id, userId, status, createdAt]
    ReservationsListResponse:
      type: object
      properties:
        reservations:
          type: array
          items:
            $ref: "#/components/schemas/Reservation"
        total:
          type: integer
      required: [reservations, total]
    CreateReservationBody:
      type: object
      properties:
        bookId:
          type: integer
        storeBookId:
          type: integer
...